Change balance dtos to balance service models
#164
