--- portfolio/src/main/java/com/jointrivial/portfolio/model/dto/BalanceDto.java
@@ -1,15 +1,0 @@
-package com.jointrivial.portfolio.model.dto;
-
-public class BalanceDto {
-
-    private BalanceAmountDto balanceAmount;
-
-    public BalanceAmountDto getBalanceAmount() {
-        return balanceAmount;
-    }
-
-    public BalanceDto setBalanceAmount(BalanceAmountDto balanceAmount) {
-        this.balanceAmount = balanceAmount;
-        return this;
-    }
-}
