--- portfolio/src/main/java/com/jointrivial/portfolio/model/dto/UserBalancesRootDto.java
@@ -1,17 +1,0 @@
-package com.jointrivial.portfolio.model.dto;
-
-import java.util.List;
-
-public class UserBalancesRootDto {
-
-    private List<BalancesRootDto> userBalances;
-
-    public List<BalancesRootDto> getUserBalances() {
-        return userBalances;
-    }
-
-    public UserBalancesRootDto setUserBalances(List<BalancesRootDto> userBalances) {
-        this.userBalances = userBalances;
-        return this;
-    }
-}
