+++ portfolio/src/main/java/com/jointrivial/portfolio/model/service/balance/BalanceServiceModel.java
@@ -1,0 +1,15 @@
+package com.jointrivial.portfolio.model.service.balance;
+
+public class BalanceServiceModel {
+
+    private BalanceAmountServiceModel balanceAmount;
+
+    public BalanceAmountServiceModel getBalanceAmount() {
+        return balanceAmount;
+    }
+
+    public BalanceServiceModel setBalanceAmount(BalanceAmountServiceModel balanceAmount) {
+        this.balanceAmount = balanceAmount;
+        return this;
+    }
+}
