+++ portfolio/src/main/java/com/jointrivial/portfolio/model/service/balance/UserBalancesRootServiceModel.java
@@ -1,0 +1,17 @@
+package com.jointrivial.portfolio.model.service.balance;
+
+import java.util.List;
+
+public class UserBalancesRootServiceModel {
+
+    private List<BalancesRootServiceModel> userBalances;
+
+    public List<BalancesRootServiceModel> getUserBalances() {
+        return userBalances;
+    }
+
+    public UserBalancesRootServiceModel setUserBalances(List<BalancesRootServiceModel> userBalances) {
+        this.userBalances = userBalances;
+        return this;
+    }
+}
