+++ portfolio/src/main/java/com/jointrivial/portfolio/service/AssetService.java
@@ -1,11 +1,11 @@
 package com.jointrivial.portfolio.service;
 
-import com.jointrivial.portfolio.model.dto.BalancesRootDto;
+import com.jointrivial.portfolio.model.service.balance.BalancesRootServiceModel;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 
 public interface AssetService {
 
-    BalancesRootDto getAllBalances(String keyOrganizations) throws URISyntaxException, IOException, InterruptedException;
+    BalancesRootServiceModel getAllBalances(String keyOrganizations) throws URISyntaxException, IOException, InterruptedException;
 }
