+++ portfolio/src/main/java/com/jointrivial/portfolio/service/impl/AssetServiceImpl.java
@@ -1,7 +1,7 @@
 package com.jointrivial.portfolio.service.impl;
 
 import com.google.gson.Gson;
-import com.jointrivial.portfolio.model.dto.BalancesRootDto;
+import com.jointrivial.portfolio.model.service.balance.BalancesRootServiceModel;
 import com.jointrivial.portfolio.service.AssetService;
 import org.springframework.stereotype.Service;
 
@@ -24,18 +24,18 @@ public class AssetServiceImpl implements AssetService {
     }
 
     @Override
-    public BalancesRootDto getAllBalances(String keyOrganizations) throws URISyntaxException, IOException, InterruptedException {
+    public BalancesRootServiceModel getAllBalances(String keyOrganization) throws URISyntaxException, IOException, InterruptedException {
 
         HttpClient client = HttpClient.newHttpClient();
 
         HttpRequest request = HttpRequest.newBuilder()
                 .uri(new URI(BALANCES_URL))
-                .header("KeyOrganizations", keyOrganizations)
+                .header("Key-Organization", keyOrganization)
                 .GET()
                 .build();
 
         HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
 
-        return gson.fromJson(response.body(), BalancesRootDto.class);
+        return gson.fromJson(response.body(), BalancesRootServiceModel.class);
     }
 }
